feat: デスクトップ更新通知をフッターの更新ボタンに変更
- 更新モーダル(UpdateModal)を廃止し、フッターに常設の更新ボタン(UpdateButton)を追加
- アウトライン基調の控えめなデザイン。文言は「再起動して更新」＋新バージョン表示
- 押下でダウンロード（ボタンが進捗表示に変化）→インストール→自動再起動
- 未使用になったi18nキー(update.title/now/later)を整理（closeはSettingsModalが使用のため維持）
- リリース時にannouncements.jsonで新機能紹介を出す運用をCLAUDE.mdに追記
