+++ CLAUDE.md
@@ -369,10 +369,37 @@ vX.X.X has been released!
 
 ⚠️ リリースノートが空のまま公開しないこと。タグ push 後、デスクトップビルド完了を待つ間にリリースノートを書く。
 
-### 7. 最終確認
+### 7. お知らせ（新機能紹介）の追加 — `announcements.json`
+
+注目機能を含むリリースでは、アプリ内お知らせ（`AnnouncementBanner`）にエントリを1件追加して新機能を紹介すること。**新機能紹介用のモーダルは作らない**（モーダルは使わない方針）。お知らせはこの帯に寄せる。
+
+- 編集対象はリポジトリ**ルートの `announcements.json`**。クライアントは **main の生ファイル**（`https://raw.githubusercontent.com/oboroge0/AITuberFlow/main/announcements.json`）を毎回取得するため、**main にマージした瞬間に全ユーザーへ配信される（ビルド不要・即本番反映）**。内容は慎重に。
+- ホーム画面上部に帯で表示され、`type` で配色（`info`=青／`warning`=黄／`critical`=赤）。× で閉じると localStorage に記憶され再表示されない。
+- 新機能紹介は `type: "info"` を使う。**title / message は日英両方**を必ず書く。
+- `targetVersions` は**完全一致の表示許可リスト**（範囲指定ではない）。
+  - 新機能を「更新済みの人にだけ」見せる → その新バージョンだけ列挙（例 `["2.5.2"]`）
+  - 全ユーザーに見せる → `targetVersions` を省略
+  - 旧版ユーザーへ「更新して」と促す → 対象の旧バージョンを列挙
+- 不要になった古いお知らせは削除してよい（`dismissedIds` は自動で掃除される）。
+
+フォーマット:
+
+```json
+{
+  "id": "feature-vX.Y.Z",
+  "type": "info",
+  "title":   { "ja": "…", "en": "…" },
+  "message": { "ja": "…", "en": "…" },
+  "targetVersions": ["X.Y.Z"],
+  "date": "YYYY-MM-DD"
+}
+```
+
+### 8. 最終確認
 
 - [ ] タグが正しいコミットを指している
 - [ ] マイルストーンのissueがクローズされている
 - [ ] CIのリリースワークフローが成功している
 - [ ] **リリースノートが日英両方で記載されている**
 - [ ] デスクトップビルドの成果物がリリースに添付されている
+- [ ] 注目機能がある場合、`announcements.json` に新機能紹介を追加した（日英両方）
