+++ apps/web/locales/ja.json
@@ -455,11 +455,9 @@
     }
   },
   "update": {
-    "title": "アップデート",
+    "restart": "再起動して更新",
     "downloading": "ダウンロード中...",
     "error": "アップデートに失敗しました",
-    "now": "今すぐ更新",
-    "later": "あとで",
     "close": "閉じる",
     "retry": "再試行"
   },
